Recover typed UI base class interfaces
Rename recovered TEventHandler, TView, TControl, and TWindow methods across their override families; promote page, geometry, dialog, and QuickDraw cursor signatures to real types; and add the geometry-type gate. Port the two retail QuickDraw cursor compatibility stubs and the concrete Windows cursor overrides.\n\nVerified with just agent-check and just precommit. Aggregate vtables are 100%. Stats: +10 exact functions, +0.10 pp average similarity, and +3 paired globals; structured triage found no actionable regression in TWindow::Show and retained evidence-backed typed source for the two nonmatching branch shapes.

+++ include/game/TDialogBehavior.h
@@ -39,10 +39,10 @@ public:
   // cancel command on Escape/Delete. ---
   unsigned char armed; // 0x10 — state/flag byte
   unsigned char padding_11_13[0x03];
-  int defaultCommandCode;       // 0x14 — command fired on Enter/Return
-  int cancelCommandCode;        // 0x18 — command fired on Escape/Delete
-  int armedCommandCode;         // 0x1c — command armed via slot 0x0e
-  unsigned char dismissPending; // 0x20 — set by Dismiss, cleared before the modal loop
+  unsigned long defaultCommandCode; // 0x14 — command fired on Enter/Return
+  unsigned long cancelCommandCode;  // 0x18 — command fired on Escape/Delete
+  unsigned long armedCommandCode;   // 0x1c — command armed via slot 0x0e
+  unsigned char dismissPending;     // 0x20 — set by Dismiss, cleared before the modal loop
   unsigned char padding_21_23[0x03];
 
   TDialogBehavior();
